Add CMake

+++ include/binsparse/matrix.h
@@ -9,9 +9,9 @@ typedef struct bsp_matrix_t {
   bsp_array_t indices_1;
   bsp_array_t pointers_to_1;
 
-  int nrows;
-  int ncols;
-  int nnz;
+  size_t nrows;
+  size_t ncols;
+  size_t nnz;
 
   bsp_matrix_format_t format;
 } bsp_matrix_t;
